% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_dist}
\alias{infer_dist}
\title{Infers cosimu distribution parameters from the log2 fold-change expression of
a real data vector.}
\usage{
infer_dist(
  lfc_vector,
  min.size = 0.01,
  sym_mode = FALSE,
  plot = FALSE,
  thresh_up = 0.1,
  thresh_down = -0.1,
  scale = 0.2,
  nr_mean = 0,
  nr_sd = 0.15,
  save_yaml = NULL,
  return = FALSE
)
}
\arguments{
\item{lfc_vector}{vector of log2 fold-change values.}

\item{min.size}{the minimum size of the modes. Default tos to 0.01, eliminating modes
with less than 1\% of the distributional area. See \code{LaplacesDemon} for more details.}

\item{sym_mode}{boolean set to \code{TRUE} to return a symmetrical parametrization
for up and down regulated modalities. Default to \code{FALSE}.}

\item{plot}{boolean set to \code{TRUE} to plot the log2 fold change distribution and
its modes, in order to get symmetrical distributions for the deregulated
modalities. Default to \code{FALSE}.}

\item{thresh_up}{threshold to filter the up-regulated modes. Default to 0.1.}

\item{thresh_down}{threshold to filter the down-regulated modes. Default to -0.1.}

\item{scale}{scale of the gamma distributions. Needs to be fixed in order to
infer the associated shapes. Default to 0.2.}

\item{nr_mean}{mean of the Gaussian distribution of the NR modality.
Default to 0.}

\item{nr_sd}{standard deviation of the Gaussian distribution of the NR
modality. Default to 0.}

\item{save_yaml}{character string naming a file to save the inferred parameters.
Default to \code{NULL}.}

\item{return}{boolean set to \code{TRUE} to return the parameter values. Note that the
quantile functions are only determined during the loading process, see \code{load_dist}.
Default to \code{FALSE}.}
}
\description{
Infers cosimu distribution parameters from the log2 fold-change expression of
a real data vector.
}
