% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dist}
\alias{load_dist}
\title{Loads the YAML file containing the distribution parameters.}
\usage{
load_dist(yaml_file)
}
\arguments{
\item{yaml_file}{character string naming a file to load the inferred parameters.}
}
\value{
list including:
\itemize{
\item prop_sm_up: Vector of proportions of the sub-modalities of the up-regulated
entities.
\item prop_sm_down: Vector of proportions of the sub-modalities of the
down-regulated entities.
\item up_means: Vector of means of the up sub-modalities
\item nr_means: Mean of the NR (sub-)modality
\item down_means: Vector of means of the down sub-modalities
\item qf_vect_up: List of quantile functions associated with each sub-modality of
the up-regulated modality.
\item qf_vect_nr: List of quantile functions associated with each sub-modality of
the non-deregulated modality.
\item qf_vect_down : List of quantile functions associated to each sub-modality
of the down-deregulated modality.
\item alpha: Fraction of deregulated genes that are up-regulated. The fraction of
deregulated genes that are down-regulated is 1-alpha. (p_up = alpha*p_deg and
p_down = (1-alpha)*p_deg with 0 < p_deg < 1).
}
}
\description{
This function loads the YAML file generated by \code{infer_dist} and prepares the
parameters, including quantile functions, for use in the cosimu function.
}
