% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_simple_param}
\alias{load_simple_param}
\title{Loads the YAML file containing the parameters needed for a simple simulation.}
\usage{
load_simple_param(yaml_file)
}
\arguments{
\item{yaml_file}{character string naming a file to load the inferred parameters.
The file must have the structure shown in the cosimu_application repository
simple_yaml_structure.yaml}
}
\value{
list of parameters
}
\description{
Loads the YAML file containing the parameters needed for a simple simulation.
}
