% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ModalityObj.R
\name{ModalityObj}
\alias{ModalityObj}
\title{Global Modality object}
\description{
Parent Class for the modality objects, defines the common parameteres
and their get functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_nb_ent}{\code{ModalityObj$get_nb_ent()}}
\item \href{#method-get_p_up}{\code{ModalityObj$get_p_up()}}
\item \href{#method-get_p_down}{\code{ModalityObj$get_p_down()}}
\item \href{#method-get_values}{\code{ModalityObj$get_values()}}
\item \href{#method-clone}{\code{ModalityObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_nb_ent"></a>}}
\if{latex}{\out{\hypertarget{method-get_nb_ent}{}}}
\subsection{Method \code{get_nb_ent()}}{
Gets the number of entities in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModalityObj$get_nb_ent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an integer equal to the number of entities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_p_up"></a>}}
\if{latex}{\out{\hypertarget{method-get_p_up}{}}}
\subsection{Method \code{get_p_up()}}{
Gets the proportion of up regulated entities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModalityObj$get_p_up()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a double equal to the proportion of up regulates
entities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_p_down"></a>}}
\if{latex}{\out{\hypertarget{method-get_p_down}{}}}
\subsection{Method \code{get_p_down()}}{
Gets the proportion of down regulated entities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModalityObj$get_p_down()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a double equal to the proportion of down regulates
entities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-get_values}{}}}
\subsection{Method \code{get_values()}}{
Gets the modalities vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModalityObj$get_values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named vector of integers (factors) corresponding to the
modality of each entity.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModalityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
