% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ModalityObj.R
\name{PrimaryModalityObj}
\alias{PrimaryModalityObj}
\title{Primary modality object}
\description{
Object used as first layer for the independent
("non connected") signature
}
\section{Super class}{
\code{\link[cosimu:ModalityObj]{cosimu::ModalityObj}} -> \code{PrimaryModalityObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PrimaryModalityObj-new}{\code{PrimaryModalityObj$new()}}
\item \href{#method-PrimaryModalityObj-clone}{\code{PrimaryModalityObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_nb_ent"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_nb_ent'><code>cosimu::ModalityObj$get_nb_ent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_p_down"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_p_down'><code>cosimu::ModalityObj$get_p_down()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_p_up"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_p_up'><code>cosimu::ModalityObj$get_p_up()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_values"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_values'><code>cosimu::ModalityObj$get_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimaryModalityObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-PrimaryModalityObj-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{PrimaryModalityObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimaryModalityObj$new(nb_ent, p_up, p_down, entity_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nb_ent}}{: Integer representing the number of entities
(genes or transcripts)}

\item{\code{p_up}}{: Proportion of up regulated entities}

\item{\code{p_down}}{: Proportion of down regulated entities}

\item{\code{entity_id}}{(optional) : Character vector with the entity names.
Default \code{NULL}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimaryModalityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimaryModalityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimaryModalityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
