% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProbabilityObj.R
\name{PrimaryProbabilityObj}
\alias{PrimaryProbabilityObj}
\title{Primary probability object}
\description{
Object used as third layer for the primary signature
}
\section{Super class}{
\code{\link[cosimu:ProbabilityObj]{cosimu::ProbabilityObj}} -> \code{PrimaryProbabilityObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PrimaryProbabilityObj-new}{\code{PrimaryProbabilityObj$new()}}
\item \href{#method-PrimaryProbabilityObj-clone}{\code{PrimaryProbabilityObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ProbabilityObj" data-id="get_values"><a href='../../cosimu/html/ProbabilityObj.html#method-ProbabilityObj-get_values'><code>cosimu::ProbabilityObj$get_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimaryProbabilityObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-PrimaryProbabilityObj-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{PrimaryProbabilityObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimaryProbabilityObj$new(nb_ent, entity_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nb_ent}}{: Integer representing the number of entities
(genes or transcripts)}

\item{\code{entity_id}}{(optional) : character vector with the entity names.
Default \code{NA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimaryProbabilityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimaryProbabilityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimaryProbabilityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
