% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SignatureObj.R
\name{PrimarySignatureObj}
\alias{PrimarySignatureObj}
\title{Primary signature object}
\description{
Object composed by three layers:
\code{PrimaryModalityObj},\code{PrimarySubmodalityObj}
and \code{PrimaryProbabilityObj}.
}
\section{Super class}{
\code{\link[cosimu:SignatureObj]{cosimu::SignatureObj}} -> \code{PrimarySignatureObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrimarySignatureObj$new()}}
\item \href{#method-clone}{\code{PrimarySignatureObj$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_base_expression">}\href{../../cosimu/html/SignatureObj.html#method-get_base_expression}{\code{cosimu::SignatureObj$get_base_expression()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_lfc_vect">}\href{../../cosimu/html/SignatureObj.html#method-get_lfc_vect}{\code{cosimu::SignatureObj$get_lfc_vect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_modality_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_modality_obj}{\code{cosimu::SignatureObj$get_modality_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_probability_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_probability_obj}{\code{cosimu::SignatureObj$get_probability_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_submodality_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_submodality_obj}{\code{cosimu::SignatureObj$get_submodality_obj()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{PrimarySignatureObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimarySignatureObj$new(
  nb_ent,
  p_up,
  p_down,
  prop_sm_up,
  prop_sm_down,
  qf_vect_up,
  qf_vect_nr,
  qf_vect_down,
  entity_id = NULL,
  base_expression = FALSE,
  initial_base = NA,
  up_means = NA,
  nr_means = NA,
  down_means = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nb_ent}}{: Integer representing the number of entities
(genes or transcripts).}

\item{\code{p_up}}{: Proportion of up-regulated entities.}

\item{\code{p_down}}{: Proportion of down-regulated entities.}

\item{\code{prop_sm_up}}{: Vector of proportions of the
sub-modalities of the up-regulated entities.}

\item{\code{prop_sm_down}}{: Vector of proportions of the
sub-modalities of the down-regulated entities.}

\item{\code{qf_vect_up}}{: List of quantile functions associated
to each sub-modality of the up-regulated modality.}

\item{\code{qf_vect_nr}}{: List of quantile functions associated
to each sub-modality of the non-deregulated modality.}

\item{\code{qf_vect_down}}{: List of quantile functions associated
to each sub-modality of the down-deregulated modality.}

\item{\code{entity_id}}{(optional) : Character vector with the entity names.}

\item{\code{base_expression}}{(optional) : Boolean set to \code{TRUE} if their is a
need to correct the base expression. Default \code{FALSE}.}

\item{\code{initial_base}}{(optional) : Double vector with the normalized base
expression values to be corrected and control simulated bias.}

\item{\code{up_means}}{(optional) : Vector of means of the up sub-modalities use
if \code{base_expression = TRUE}, to apply the base expression correction.}

\item{\code{nr_means}}{(optional) : Vector with the mean of the non-deregulated
use if \code{base_expression = TRUE}, to apply the base expression correction.}

\item{\code{down_means}}{(optional) : Vector of means of the down sub-modalities use
if \code{base_expression = TRUE}, to apply the base expression correction.
Default \code{NA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimarySignatureObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
