% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProbabilityObj.R
\name{ProbabilityObj}
\alias{ProbabilityObj}
\title{Global Probability object}
\description{
Global Probability object

Global Probability object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProbabilityObj-get_values}{\code{ProbabilityObj$get_values()}}
\item \href{#method-ProbabilityObj-clone}{\code{ProbabilityObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbabilityObj-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-ProbabilityObj-get_values}{}}}
\subsection{Method \code{get_values()}}{
Gets the a probability vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbabilityObj$get_values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named vector of integers (factors) corresponding to the
probability of each entity.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbabilityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProbabilityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbabilityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
