% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_counts_gen}
\alias{read_counts_gen}
\title{Generation of the read counts from the fold-change signatures. This function
was inspired from the first part of the function \code{simulate_experiment} from the
package newpolyester.
(Frazee AC, Jaffe AE, Langmead B, Leek JT.
Polyester: simulating RNA-seq datasets with differential transcript expression.
Bioinformatics. 2015 Sep 1;31(17):2778-84. doi: 10.1093/bioinformatics/btv272.
Epub 2015 Apr 28. PMID: 25926345; PMCID: PMC4635655.)}
\usage{
read_counts_gen(
  num_reps = c(3, 3),
  p_nb_mol = 300,
  size = NULL,
  size_factor = 3,
  fold_changes,
  num_reads = 1e+06,
  TP_analysis = FALSE,
  length_ent = NULL,
  nb_ent = NULL,
  names_ent = NULL,
  seed = NULL
)
}
\arguments{
\item{num_reps}{number of replicates for each condition}

\item{p_nb_mol}{proportional to true number of molecules associated to each
gene in its a basal expression level (untreated) (inspired from DESeq2
proportional to true number of fragments, default 300)}

\item{size}{optional size for the NB model}

\item{size_factor}{optional factor that set the size of the NB to mean/size_factor}

\item{fold_changes}{fold-changes values}

\item{num_reads}{expected number of reads in each experiment}

\item{TP_analysis}{boolean set to TRUE if a 3'-RNASeq analysis parametrization
is considered. This impacts the influence of the gene or transcripts length
in the analysis, that is not take into account in a 3' analysis.}

\item{length_ent}{gene or transcripts lengths.}

\item{nb_ent}{number of genes or transcripts.}

\item{names_ent}{names of the genes or transcripts.}

\item{seed}{random seed.}
}
\description{
Generation of the read counts from the fold-change signatures. This function
was inspired from the first part of the function \code{simulate_experiment} from the
package newpolyester.
(Frazee AC, Jaffe AE, Langmead B, Leek JT.
Polyester: simulating RNA-seq datasets with differential transcript expression.
Bioinformatics. 2015 Sep 1;31(17):2778-84. doi: 10.1093/bioinformatics/btv272.
Epub 2015 Apr 28. PMID: 25926345; PMCID: PMC4635655.)
}
