% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ModalityObj.R
\name{SecondaryModalityObj}
\alias{SecondaryModalityObj}
\title{Secondary modality object}
\description{
used as first layer for the connected signature
}
\section{Super class}{
\code{\link[cosimu:ModalityObj]{cosimu::ModalityObj}} -> \code{SecondaryModalityObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondaryModalityObj-new}{\code{SecondaryModalityObj$new()}}
\item \href{#method-SecondaryModalityObj-get_transition_mat}{\code{SecondaryModalityObj$get_transition_mat()}}
\item \href{#method-SecondaryModalityObj-clone}{\code{SecondaryModalityObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_nb_ent"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_nb_ent'><code>cosimu::ModalityObj$get_nb_ent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_p_down"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_p_down'><code>cosimu::ModalityObj$get_p_down()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_p_up"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_p_up'><code>cosimu::ModalityObj$get_p_up()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ModalityObj" data-id="get_values"><a href='../../cosimu/html/ModalityObj.html#method-ModalityObj-get_values'><code>cosimu::ModalityObj$get_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryModalityObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryModalityObj-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{SecondaryModalityObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryModalityObj$new(
  prim_mod_obj,
  connectivity_score = NA,
  mod_transition = c("default", "customized"),
  transition_mat = NA,
  nr_noise = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prim_mod_obj}}{: \code{PrimaryModalityObj} used as base for the
simulation of the connected object.}

\item{\code{connectivity_score}}{(optional) : Double between -1 and 1
representing the "connectivity" between the two perturbations.
Used to generate the transition matrix if the modality transition is not
"customized".}

\item{\code{mod_transition}}{: Modality transition "default"
or "customized".}

\item{\code{transition_mat}}{(optional) : Transition matrix used if
\code{mod_transition} = "customized". Default \code{NA}.}

\item{\code{nr_noise}}{(optional) : Double between 0 and 0.5 representing the
noise induced by the non-deregulated entities during the "default"
transition. Default \code{NA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryModalityObj-get_transition_mat"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryModalityObj-get_transition_mat}{}}}
\subsection{Method \code{get_transition_mat()}}{
Get the transition matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryModalityObj$get_transition_mat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
transition matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryModalityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryModalityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryModalityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
