% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProbabilityObj.R
\name{SecondaryProbabilityObj}
\alias{SecondaryProbabilityObj}
\title{Secondary probability object}
\description{
Object used as third layer for the connected signature
}
\section{Super class}{
\code{\link[cosimu:ProbabilityObj]{cosimu::ProbabilityObj}} -> \code{SecondaryProbabilityObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondaryProbabilityObj-new}{\code{SecondaryProbabilityObj$new()}}
\item \href{#method-SecondaryProbabilityObj-clone}{\code{SecondaryProbabilityObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="ProbabilityObj" data-id="get_values"><a href='../../cosimu/html/ProbabilityObj.html#method-ProbabilityObj-get_values'><code>cosimu::ProbabilityObj$get_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryProbabilityObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryProbabilityObj-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{SecondaryProbabilityObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryProbabilityObj$new(
  prim_perc_obj,
  proba_transition,
  copula,
  theta,
  nbins,
  ncpus
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prim_perc_obj}}{: \code{PrimaryProbabilityObj} used as base for the
simulation of the connected object.}

\item{\code{proba_transition}}{: choice of the probability transition
"independent", "cop" or "deterministic".}

\item{\code{copula}}{(optional) : 'Gauss','Plackett' or 'Frank' copula
used when the probability transition is "cop".}

\item{\code{theta}}{(optional): parameter of the copula for the "cop"
transition. If:
\itemize{
\item cop = "Frank": theta \verb{[-Inf, Inf]}; theta → Inf comonotonicity ;
theta → -Inf countermonotonicity; theta → 0 independence copula.
\item cop = "Plackett": theta (0, Inf) \ {1}; theta → Inf comonotonicity ;
theta → 0 countermonotonicity; theta → 1 independence copula.
\item cop = "Gauss": theta \verb{[-1,1]} ; theta → 1 comonotonicity ;
theta → -1 countermonotonicity; theta → 0 independence copula
}}

\item{\code{nbins}}{(optional): number of bins used to generate the
probability layer by the probability "cop" transition.}

\item{\code{ncpus}}{(optional) : number of cpus for parallel calculation used in
'cop' probability transition.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryProbabilityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryProbabilityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryProbabilityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
