% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SignatureObj.R
\name{SecondarySignatureObj}
\alias{SecondarySignatureObj}
\title{Secondary signature object}
\description{
Object composed by three layers:
\code{SecondaryModalityObj},\code{SecondarySubmodalityObj}
and \code{SecondaryProbabilityObj}.
}
\section{Super class}{
\code{\link[cosimu:SignatureObj]{cosimu::SignatureObj}} -> \code{SecondarySignatureObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SecondarySignatureObj$new()}}
\item \href{#method-get_connectivities}{\code{SecondarySignatureObj$get_connectivities()}}
\item \href{#method-clone}{\code{SecondarySignatureObj$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_base_expression">}\href{../../cosimu/html/SignatureObj.html#method-get_base_expression}{\code{cosimu::SignatureObj$get_base_expression()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_lfc_vect">}\href{../../cosimu/html/SignatureObj.html#method-get_lfc_vect}{\code{cosimu::SignatureObj$get_lfc_vect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_modality_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_modality_obj}{\code{cosimu::SignatureObj$get_modality_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_probability_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_probability_obj}{\code{cosimu::SignatureObj$get_probability_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cosimu" data-topic="SignatureObj" data-id="get_submodality_obj">}\href{../../cosimu/html/SignatureObj.html#method-get_submodality_obj}{\code{cosimu::SignatureObj$get_submodality_obj()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{SecondarySignatureObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySignatureObj$new(
  prim_sig_obj,
  mod_transition,
  submod_transition,
  proba_transition,
  qf_vect_up = NA,
  qf_vect_nr = NA,
  qf_vect_down = NA,
  connectivity_score = NA,
  transition_mat = NA,
  nr_noise = NA,
  prop_sm_up = NA,
  prop_sm_down = NA,
  rho_submod = NA,
  copula_submod = NA,
  eps_submod = NA,
  optim_method_submod = "Brent",
  copula_prob = "Frank",
  theta_prob = 10,
  nbins_prob = 1000,
  base_expression = FALSE,
  up_means = NA,
  nr_means = NA,
  down_means = NA,
  ncpus = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prim_sig_obj}}{: \code{PrimarySignatureObj} used as base for the
simulation of the connected object.}

\item{\code{mod_transition}}{: Modality transition "default"
or "customized".}

\item{\code{submod_transition}}{: Sub-modality transition
"cop", "independent", "stochastic" or "deterministic".}

\item{\code{proba_transition}}{: Probability transition
"independent", "stochastic" or "deterministic".}

\item{\code{qf_vect_up}}{(optional) : list of quantile functions associated
to each sub-modality of the up-regulated modality. Default \code{NA}.}

\item{\code{qf_vect_nr}}{(optional) : list of quantile functions associated
to each sub-modality of the non-deregulated modality. Default \code{NA}.}

\item{\code{qf_vect_down}}{(optional) : list of quantile functions associated
to each sub-modality of the down-deregulated modality. Default \code{NA}.}

\item{\code{connectivity_score}}{(optional) : Double between -1 and 1
representing the "connectivity" between the two perturbations.
Used to generate the transition matrix if the modality transition is not
"customized".}

\item{\code{transition_mat}}{(optional) : transition matrix used if
\code{mod_transition} = "customized". Default \code{NA}.}

\item{\code{nr_noise}}{(optional) : Double between 0 and 0.5 representing the
noise induced by the non-deregulated entities during the "default"
transition. Default \code{NA}.}

\item{\code{prop_sm_up}}{(optional) : vector of proportions of the
sub-modalities of the up-regulated entities. Default \code{NA}.}

\item{\code{prop_sm_down}}{(optional) : vector of proportions of the
sub-modalities of the down-deregulated entities. Default \code{NA}.}

\item{\code{rho_submod}}{(optional) : correlation to be reached by the
sub-modality "cop" transition.}

\item{\code{copula_submod}}{(optional) : 'Gauss','Plackett' or 'Frank' copula
used when the sub-modality transition is "cop". Default \code{NA}.}

\item{\code{eps_submod}}{(optional) : Error between the expected correlation
(\code{rho_submod}) and the one estimated with the copula function.
Default \code{NA}.}

\item{\code{optim_method_submod}}{(optional) : Optimization method used
during the sub-modality "cop" transition. See \code{optim} for more
details. Default "Brent".}

\item{\code{copula_prob}}{(optional) : 'Gauss','Plackett' or 'Frank' copula
used when the probability transition is "cop". Default \code{Frank}.}

\item{\code{theta_prob}}{: parameter of the copula for the "cop" transition. If:
\itemize{
\item cop = "Frank": theta \link{-Inf, Inf}; theta → Inf comonotonicity ;
theta → -Inf countermonotonicity; theta → 0 independence copula.
\item cop = "Plackett": theta (0, Inf) \ {1}; theta → Inf comonotonicity ;
theta → 0 countermonotonicity; theta → 1 independence copula.
\item cop = "Gauss": theta \link{-1,1} ; theta → 1 comonotonicity ;
theta → -1 countermonotonicity; theta → 0 independence copula
Default 10.
}}

\item{\code{nbins_prob}}{(optional): number of bins used to generate the
probability layer by the probability "cop" transition. Default 1e3.}

\item{\code{base_expression}}{(optional) : Boolean set to \code{TRUE} if their is a
need to correct the base expression. Default \code{FALSE}.}

\item{\code{up_means}}{(optional) : vector of means of the up sub-modalities use
if \code{base_expression = TRUE}, to apply the base expression correction.}

\item{\code{nr_means}}{(optional) : vector with the mean of the non-deregulated
use if \code{base_expression = TRUE}, to apply the base expression correction.}

\item{\code{down_means}}{(optional) : vector of means of the down sub-modalities
use if \code{base_expression = TRUE}, to apply the base expression correction.}

\item{\code{ncpus}}{(optional) : number of cpus for parallel calculation used in
'cop' probability transition. Default 1.}

\item{\code{entity_id}}{(optional) : character vector with the entity names.
Default \code{NULL}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connectivities"></a>}}
\if{latex}{\out{\hypertarget{method-get_connectivities}{}}}
\subsection{Method \code{get_connectivities()}}{
Get the vector with the expected and the effective connectivity scores.
For all transitions different to "cop" they should be identical.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySignatureObj$get_connectivities()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with connectivities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySignatureObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
