% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SubmodalityObj.R
\name{SecondarySubmodalityObj}
\alias{SecondarySubmodalityObj}
\title{Secondary sub-modality object}
\description{
Object used as second layer for the connected signature
}
\section{Super class}{
\code{\link[cosimu:SubmodalityObj]{cosimu::SubmodalityObj}} -> \code{SecondarySubmodalityObj}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondarySubmodalityObj-new}{\code{SecondarySubmodalityObj$new()}}
\item \href{#method-SecondarySubmodalityObj-get_optim_res}{\code{SecondarySubmodalityObj$get_optim_res()}}
\item \href{#method-SecondarySubmodalityObj-get_rho}{\code{SecondarySubmodalityObj$get_rho()}}
\item \href{#method-SecondarySubmodalityObj-get_list_transition_mat}{\code{SecondarySubmodalityObj$get_list_transition_mat()}}
\item \href{#method-SecondarySubmodalityObj-clone}{\code{SecondarySubmodalityObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="SubmodalityObj" data-id="get_nb_submod_by_mod"><a href='../../cosimu/html/SubmodalityObj.html#method-SubmodalityObj-get_nb_submod_by_mod'><code>cosimu::SubmodalityObj$get_nb_submod_by_mod()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="SubmodalityObj" data-id="get_prop_list"><a href='../../cosimu/html/SubmodalityObj.html#method-SubmodalityObj-get_prop_list'><code>cosimu::SubmodalityObj$get_prop_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cosimu" data-topic="SubmodalityObj" data-id="get_values"><a href='../../cosimu/html/SubmodalityObj.html#method-SubmodalityObj-get_values'><code>cosimu::SubmodalityObj$get_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondarySubmodalityObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondarySubmodalityObj-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{SecondarySubmodalityObj} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySubmodalityObj$new(
  mod_obj,
  prim_mod_obj,
  prim_submod_obj,
  submod_transition,
  prop_sm_up = NA,
  prop_sm_down = NA,
  rho = NA,
  copula = "Frank",
  eps = 0.001,
  optim_method = "Brent"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod_obj}}{: \code{ModalityObjSecondary} corresponding to the first
layer of the connected signature.}

\item{\code{prim_mod_obj}}{: \code{ModalityObjPrimary} used as base for the
simulation of the connected object.}

\item{\code{prim_submod_obj}}{: \code{PrimarySubmodalityObj} used as base for the
simulation of the connected object.}

\item{\code{submod_transition}}{: Choice of the sub-modality transition
"cop", "independent", "stochastic" or "deterministic".}

\item{\code{prop_sm_up}}{(optional) : Vector of proportions of the
sub-modalities of the up regulated entities. Default \code{NA}.}

\item{\code{prop_sm_down}}{(optional) : Vector of proportions of the
sub-modalities of the down regulated entities. Default \code{NA}.}

\item{\code{rho}}{(optional) : Correlation to be reached by the
sub-modality "cop" transition.}

\item{\code{copula}}{(optional) : 'Gauss','Plackett' or 'Frank' copula
used when the sub-modality transition is "cop". Default "Frank".}

\item{\code{eps}}{(optional) : Error between the expected correlation
(\code{rho}) and the one estimated with the copula function.
Default "1e-3.}

\item{\code{optim_method}}{(optional) : Optimization method used
during the sub-modality "cop" transition. See \code{optim} for more
details. Default "Brent".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondarySubmodalityObj-get_optim_res"></a>}}
\if{latex}{\out{\hypertarget{method-SecondarySubmodalityObj-get_optim_res}{}}}
\subsection{Method \code{get_optim_res()}}{
Get the optimization results. See \code{optim} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySubmodalityObj$get_optim_res()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with optimization results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondarySubmodalityObj-get_rho"></a>}}
\if{latex}{\out{\hypertarget{method-SecondarySubmodalityObj-get_rho}{}}}
\subsection{Method \code{get_rho()}}{
Get the vector with the expected and the effective correlation (rho)
used for "cop" transition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySubmodalityObj$get_rho()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list with two elements expected rho and effective one
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondarySubmodalityObj-get_list_transition_mat"></a>}}
\if{latex}{\out{\hypertarget{method-SecondarySubmodalityObj-get_list_transition_mat}{}}}
\subsection{Method \code{get_list_transition_mat()}}{
Get a list of the transition matrix between all the sub-modalities
of two modalities.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySubmodalityObj$get_list_transition_mat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of transition matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondarySubmodalityObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondarySubmodalityObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondarySubmodalityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
