% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SignatureObj.R
\name{SignatureObj}
\alias{SignatureObj}
\title{Global Signature object}
\description{
Global Signature object

Global Signature object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignatureObj-get_modality_obj}{\code{SignatureObj$get_modality_obj()}}
\item \href{#method-SignatureObj-get_submodality_obj}{\code{SignatureObj$get_submodality_obj()}}
\item \href{#method-SignatureObj-get_probability_obj}{\code{SignatureObj$get_probability_obj()}}
\item \href{#method-SignatureObj-get_lfc_vect}{\code{SignatureObj$get_lfc_vect()}}
\item \href{#method-SignatureObj-get_base_expression}{\code{SignatureObj$get_base_expression()}}
\item \href{#method-SignatureObj-clone}{\code{SignatureObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-get_modality_obj"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-get_modality_obj}{}}}
\subsection{Method \code{get_modality_obj()}}{
Gets the modality object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$get_modality_obj()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{ModalityObj} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-get_submodality_obj"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-get_submodality_obj}{}}}
\subsection{Method \code{get_submodality_obj()}}{
Gets the sub-modality object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$get_submodality_obj()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{SubmodalityObj} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-get_probability_obj"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-get_probability_obj}{}}}
\subsection{Method \code{get_probability_obj()}}{
Gets the probability object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$get_probability_obj()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{SubmodalityObj} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-get_lfc_vect"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-get_lfc_vect}{}}}
\subsection{Method \code{get_lfc_vect()}}{
Gets the log2 fold change vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$get_lfc_vect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
log2 fold change vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-get_base_expression"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-get_base_expression}{}}}
\subsection{Method \code{get_base_expression()}}{
Gets the base expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$get_base_expression()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
base_expression vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignatureObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SignatureObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignatureObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
