% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_SubmodalityObj.R
\name{SubmodalityObj}
\alias{SubmodalityObj}
\title{Global sub-modality object
Parent Class for the suubmodality objects, defines the common parameteres
and their get functions.
Note : NR modality is supposed to have only one submodality.}
\description{
Global sub-modality object
Parent Class for the suubmodality objects, defines the common parameteres
and their get functions.
Note : NR modality is supposed to have only one submodality.

Global sub-modality object
Parent Class for the suubmodality objects, defines the common parameteres
and their get functions.
Note : NR modality is supposed to have only one submodality.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_prop_list}{\code{SubmodalityObj$get_prop_list()}}
\item \href{#method-get_values}{\code{SubmodalityObj$get_values()}}
\item \href{#method-get_nb_submod_by_mod}{\code{SubmodalityObj$get_nb_submod_by_mod()}}
\item \href{#method-clone}{\code{SubmodalityObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prop_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_prop_list}{}}}
\subsection{Method \code{get_prop_list()}}{
Gets the list of the  sub-modalities proportions of a specific modality
(or each modality)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmodalityObj$get_prop_list(mod = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{(optional) : Modality for which the  sub-modalities proportions
should be returned. Default 'NA' returns the vector for all
the modalities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sub-modalities proportions vector or list of all the sub-modalities
propotions vectors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-get_values}{}}}
\subsection{Method \code{get_values()}}{
Gets the a  sub-modalities vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmodalityObj$get_values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named vector of integers (factors) corresponding to the
sub-modality of each entity.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_nb_submod_by_mod"></a>}}
\if{latex}{\out{\hypertarget{method-get_nb_submod_by_mod}{}}}
\subsection{Method \code{get_nb_submod_by_mod()}}{
Gets the number of  sub-modalities in a specific modality
(or each modality)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmodalityObj$get_nb_submod_by_mod(mod = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{(optional) : Modality for which the number of sub-modalities
should be returned. Default 'NA' returns the value for all
the modalities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an integer equal to the number of entities or a vector of values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmodalityObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
